Move paste handling to an util.
